+++ trunk/castor-2002/castor/src/examples/myapp/Product.java
@@ -16,7 +16,7 @@ public class Product
     public String    name = "";
 
 
-    public double    price;
+    public float     price;
 
 
     public ProductGroup     group;
